modification du fichier data pour inclure les decimal dans le prix

+++ data/vinodb.sql
@@ -34,7 +34,7 @@ CREATE TABLE `vino__bouteille` (
   `code_saq` varchar(50) DEFAULT NULL,
   `pays` varchar(50) DEFAULT NULL,
   `description` varchar(200) DEFAULT NULL,
-  `prix_saq` float DEFAULT NULL,
+  `prix_saq` DECIMAL(10, 2) DEFAULT NULL,
   `url_saq` varchar(200) DEFAULT NULL,
   `url_img` varchar(200) DEFAULT NULL,
   `format` varchar(20) DEFAULT NULL,
